argocd_sale: subscription integration
