+++ argocd_sale/models/__init__.py
@@ -3,3 +3,4 @@ from . import account_move
 from . import application
 from . import application_template
 from . import res_partner
+from . import subscription
